+++ eNMS/models/automation.py
@@ -74,9 +74,6 @@ class Service(AbstractBase):
     pools = relationship(
         "Pool", secondary=db.service_pool_table, back_populates="services"
     )
-    superworkflow_id = db.Column(Integer, ForeignKey("service.id"))
-    superworkflow = relationship("Service", remote_side=[id])
-    superworkflow_targets = db.Column(db.SmallString, default="placeholder")
     update_pools = db.Column(Boolean, default=False)
     send_notification = db.Column(Boolean, default=False)
     send_notification_method = db.Column(db.SmallString, default="mail")